requirements.txt
GitPython>=3.1.0
click>=8.0.0
cryptography>=3.4.0
PyYAML>=6.0